change where the ConditionalField is defined
